update project to hibernate

+++ module 3/gift-certificate-system/model/src/main/java/com/epam/esm/entity/Pagination.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Pagination {
-    private long limit;
-    private long offset;
+    private int limit;
+    private int offset;
 }
